Made bug fixes.
Follow_trajectory will no longer stop if another path can be proposed in time

The multiDOFtrajectory message was slightly simplified by removing the unused "append" field.

MotionPlanner and FutureCollisionDetector now have their own unique callback queues.

Follow trajectory will now consider the time needed to decelerate to a stop when considering when to stop the drone in the case of a future collision